requirements.txt
asgiref==3.7.2
black==22.1.0
blue==0.9.1
click==8.1.5
Django==4.2.3
django-debug-toolbar==4.1.0
django-pdb==0.6.2
djangorestframework==3.14.0
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==1.0.0
pathspec==0.11.1
platformdirs==3.9.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
